requirements.txt
amqp==2.6.1
appdirs==1.4.3
asgiref==3.2.10
billiard==3.6.3.0
CacheControl==0.12.6
celery==4.4.7
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.1
django-admin-rangefilter==0.6.2
html5lib==1.0.1
idna==2.8
importlib-metadata==1.7.0
ipaddr==2.2.0
kombu==4.6.11
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2==2.8.6
pycryptodome==3.9.8
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
zipp==3.1.0
